requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
asyncio==3.4.3
atlastk==0.13.4
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iso8601==2.1.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.4.2
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
magic-filter==1.0.12
multidict==6.4.3
openai==1.77.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pypika-tortoise==0.5.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tortoise==0.1.1
tortoise-orm==0.25.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
